renaming and organizing code
