+++ src/repositories/RepositoryLocalPure.ts
@@ -1,0 +1,28 @@
+import { iClaimEdge, ClaimEdge } from "../dataModels/ClaimEdge";
+import { iAction } from "../dataModels/Action";
+import { iScore } from "../dataModels/Score";
+import { RsData, iRsData } from "../dataModels/RsData";
+import { iRepository } from "../dataModels/iRepository";
+import { iClaim } from "../dataModels/Claim";
+import { claims } from "./reducers/claims";
+import { claimEdges } from "./reducers/claimEdges";
+import { RepositoryLocalBase } from "./RepositoryLocalBase";
+
+
+export class RepositoryLocalPure extends RepositoryLocalBase implements iRepository {
+
+    constructor(
+        public rsData: iRsData = new RsData()
+    ) {
+        super(rsData);
+    }
+
+    async notify(actions: iAction[]) {
+        for (const action of actions) {
+            //TODO: add more reducers
+            this.rsData = claims(this.rsData, action);
+            this.rsData = claimEdges(this.rsData, action);
+        }
+    }
+
+}
