+++ src/repositories/reducers/actions.ts
@@ -1,5 +1,5 @@
-import { iAction } from "../dataModels/Action"
-import { iRsData } from "../dataModels/RsData"
+import { iAction } from "../../dataModels/Action"
+import { iRsData } from "../../dataModels/RsData"
 
 export function actions(state: iRsData, action: iAction[]): iRsData {
     return {
