+++ src/repositories/reducers/claims.test.skip.ts
@@ -1,7 +1,7 @@
 import { claims } from "./claims";
-import { RsData } from "../dataModels/RsData";
-import { Action } from "../dataModels/Action";
-import { Claim } from "../dataModels/Claim";
+import { RsData } from "../../dataModels/RsData";
+import { Action } from "../../dataModels/Action";
+import { Claim } from "../../dataModels/Claim";
 
 test.only('add claim to claims', () => {
     const rsData = new RsData();
